src/requirements.txt
streamlit
python-dotenv
sentence-transformers
faiss-cpu
numpy
pdfplumber
google-generativeai
requests
google-genai
